# Speed up `brew tap` for no arguments