--- a/Library/Homebrew/brew.sh
+++ b/Library/Homebrew/brew.sh
@@ -172,6 +172,11 @@ case "$@" in
     source "${HOMEBREW_LIBRARY}/Homebrew/list.sh"
     homebrew-list "$@" && exit 0
     ;;
+  # falls back to cmd/tap.rb on a non-zero return
+  tap*)
+    source "${HOMEBREW_LIBRARY}/Homebrew/tap.sh"
+    homebrew-tap "$@" && exit 0
+    ;;
   # falls back to cmd/help.rb on a non-zero return
   help | --help | -h | --usage | "-?" | "")
     homebrew-help "$@" && exit 0