--- a/Library/Homebrew/brew.sh
+++ b/Library/Homebrew/brew.sh
@@ -172,7 +172,6 @@ case "$@" in
     source "${HOMEBREW_LIBRARY}/Homebrew/list.sh"
     homebrew-list "$@" && exit 0
     ;;
-  # falls back to cmd/tap.rb on a non-zero return
   tap)
     source "${HOMEBREW_LIBRARY}/Homebrew/tap.sh"
     homebrew-tap "$@" && exit 0